+++ src/geometries/internal/susceptibility/susceptibility.jl
@@ -1,12 +1,12 @@
 module Susceptibility
 include("base.jl")
-include("parent.jl")
+include("grid.jl")
 include("cylinder.jl")
 include("annulus.jl")
 include("triangle.jl")
 
 import .Base: BaseSusceptibility, single_susceptibility, single_susceptibility_gradient
-import .Parent: FixedSusceptibility, ParentSusceptibility, susceptibility_off_resonance, off_resonance_gradient, ShiftedSusceptibility
+import .Grid: FixedSusceptibility, SusceptibilityGrid, SusceptibilityGridElement, susceptibility_off_resonance, off_resonance_gradient, dipole_approximation, dipole_approximation_repeat, SusceptibilityGridNoRepeat, SusceptibilityGridRepeat
 import .Cylinder: CylinderSusceptibility
 import .Annulus: AnnulusSusceptibility
 import .Triangle: TriangleSusceptibility